+++ CHANGELOG.md
@@ -11,3 +11,23 @@ Example:
 
 doc = pyloo.Document(autostart=True)
 
+## [0.0.3] - 2016-01-10
+### Changed
+Some class  methods are changed to be as properties via decorator
+``@property``.
+
+Old code:
+
+doc.sheets().insert_spreadsheet("Test1", 0)
+
+New code:
+
+doc.sheets.insert_spreadsheet("Test1", 0)
+
+Changed methods:
+
+Document.o_doc() -> Document.o_doc
+Document.sheets() -> Document.sheets
+Document.fields() -> Document.fields
+Fields.document() -> Fields.document
+Sheets.o_sheets() -> Sheets.o_sheets